backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
requests==2.32.3
google-genai==1.47.0
python-dotenv==1.1.0
